Add Processor trait that consumers of the library can implement to inject their message processing logic
