+++ src/lib.rs
@@ -1,4 +1,10 @@
-pub(crate) mod messages;
-mod processor;
+extern crate core;
 
-pub use processor::run;
+pub(crate) mod messages;
+pub(crate) mod processor;
+pub(crate) mod responses;
+mod runner;
+
+pub use messages::Record;
+pub use processor::Processor;
+pub use runner::run;
